+++ src/App.jsx
@@ -1,6 +1,6 @@
 import {BrowserRouter, Route, Routes} from 'react-router-dom';
-import Login from './components/Login/index.jsx';
-import Signup from './components/Signup/index.jsx';
+import Login from './components/Login/Login.jsx';
+import Signup from './components/Signup/Signup.jsx';
 
 function App() {
   return (
